# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LMS API
  version: 0.1.0
  description: OpenAPI generado manualmente basado en comportamiento actual.
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
      required: [message]
      example:
        message: Registro no encontrado.
...
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
      required: [page, limit, total, total_pages]
      example:
        page: 1
        limit: 20
        total: 1
        total_pages: 1
...
    SedeJornada:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        is_active: { type: boolean }
      required: [id, name, is_active]
      example:
        id: 1
        name: Sede A
        is_active: true
...
    Periodo:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
      required: [id, name]
      example:
        id: 1
        name: 2026-1
        start_date: '2026-01-15'
        end_date: '2026-06-30'
    Carrera:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        is_active: { type: boolean }
      required: [id, name, is_active]
      example:
        id: 1
        name: Ingenieria
        is_active: true
    Asignatura:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        is_active: { type: boolean }
      required: [id, name, is_active]
      example:
        id: 1
        name: Programacion I
        is_active: true
    Curso:
      type: object
      properties:
        id: { type: integer, example: 16 }
        name: { type: string, example: Curso 1 }
        capacity: { type: integer, nullable: true, example: 30 }
        start_date: { type: string, format: date, nullable: true }
        end_date: { type: string, format: date, nullable: true }
        periodo:
          allOf:
            - $ref: '#/components/schemas/Periodo'
        teacher:
          type: object
          properties:
            id: { type: integer }
            email: { type: string }
            first_name: { type: string }
            last_name: { type: string }
        sede_jornada:
          allOf:
            - $ref: '#/components/schemas/SedeJornada'
        carrera:
          allOf:
            - $ref: '#/components/schemas/Carrera'
        asignatura:
          allOf:
            - $ref: '#/components/schemas/Asignatura'
      required: [id, name]
      example:
        id: 1
        name: Curso 1
        capacity: 30
        start_date: '2026-02-01'
        end_date: '2026-06-01'
        periodo:
          id: 1
          name: 2026-1
          start_date: '2026-01-15'
          end_date: '2026-06-30'
        teacher:
          id: 2
          email: teacher@lms.local
          first_name: Teacher
          last_name: One
        sede_jornada:
          id: 1
          name: Sede A
          is_active: true
        carrera:
          id: 1
          name: Ingenieria
          is_active: true
        asignatura:
          id: 1
          name: Programacion I
          is_active: true
    CursoVirtual:
      type: object
      properties:
        id: { type: integer }
        description: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        curso:
          type: object
          properties:
            id: { type: integer }
            name: { type: string }
      required: [id, created_at, curso]
      example:
        id: 1
        description: Desc
        created_at: '2026-02-05 12:00:00'
        curso:
          id: 1
          name: Curso 1
...
    FileObject:
      type: object
      properties:
        id: { type: integer }
        key: { type: string }
        bucket: { type: string }
        original_name: { type: string }
        mime_type: { type: string }
        size: { type: integer }
      required: [id, key, bucket, original_name, mime_type, size]
      example:
        id: 10
        key: uploads/2026/02/abc.txt
        bucket: lms
        original_name: test.txt
        mime_type: text/plain
        size: 5
...
paths:
...
  /structure/cursos:
    get:
      tags: [Structure]
      security: [bearerAuth: []]
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: periodo_id
          schema: { type: integer }
        - in: query
          name: teacher_id
          schema: { type: integer }
        - in: query
          name: sede_jornada_id
          schema: { type: integer }
        - in: query
          name: carrera_id
          schema: { type: integer }
        - in: query
          name: asignatura_id
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Curso' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
    post:
      tags: [Structure]
      security: [bearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 150 }
                capacity: { type: integer, minimum: 0 }
                start_date: { type: string, format: date }
                end_date: { type: string, format: date }
                periodo_id: { type: integer }
                teacher_id: { type: integer }
                sede_jornada_id: { type: integer }
                carrera_id: { type: integer }
                asignatura_id: { type: integer }
              required: [name]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Curso' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
  /structure/cursos/{id}:
    put:
      tags: [Structure]
      security: [bearerAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 150 }
                capacity: { type: integer, minimum: 0 }
                start_date: { type: string, format: date }
                end_date: { type: string, format: date }
                periodo_id: { type: integer }
                teacher_id: { type: integer }
                sede_jornada_id: { type: integer }
                carrera_id: { type: integer }
                asignatura_id: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Curso' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
    delete:
      tags: [Structure]
      security: [bearerAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
        '409':
          description: Conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
  /virtual/cursos:
    get:
      tags: [Virtual]
      security: [bearerAuth: []]
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: curso_id
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CursoVirtual' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
    post:
      tags: [Virtual]
      security: [bearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                curso_id: { type: integer }
                description: { type: string, maxLength: 2000 }
              required: [curso_id]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/CursoVirtual' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
  /virtual/cursos/{id}:
    put:
      tags: [Virtual]
      security: [bearerAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                curso_id: { type: integer }
                description: { type: string, maxLength: 2000 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/CursoVirtual' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
    delete:
      tags: [Virtual]
      security: [bearerAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
        '409':
          description: Conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
...
  /files/presign:
    post:
      tags: [Files]
      security: [bearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename: { type: string }
                mime_type: { type: string }
                size: { type: integer }
              required: [filename, mime_type, size]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      url: { type: string }
                      key: { type: string }
                      bucket: { type: string }
                      expires_in: { type: integer }
  /files/complete:
    post:
      tags: [Files]
      security: [bearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key: { type: string }
                bucket: { type: string }
                original_name: { type: string }
                mime_type: { type: string }
                size: { type: integer }
              required: [key, bucket, original_name, mime_type, size]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/FileObject' }
  /files/{id}:
    get:
      tags: [Files]
      security: [bearerAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/FileObject' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
  /files/{id}/download:
    get:
      tags: [Files]
      security: [bearerAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: query
          name: disposition
          schema: { type: string, enum: [attachment, inline] }
        - in: query
          name: filename
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      url: { type: string }
                      key: { type: string }
                      bucket: { type: string }
                      expires_in: { type: integer }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
  /files/{id}/stream:
    get:
      tags: [Files]
      security: [bearerAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: query
          name: disposition
          schema: { type: string, enum: [attachment, inline] }
        - in: query
          name: filename
          schema: { type: string }
      responses:
        '200':
          description: Streamed file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
...